decentralized-memory-marketplace/discover/discover_enhanced_agents.py
"""
Enhanced Agent Discovery with Chat Protocol Support
Discovers and tests communication with Memory NFT Marketplace agents using chat protocol.
"""

import asyncio
import requests
import json
from datetime import datetime
import os
from dotenv import load_dotenv

load_dotenv()

AGENTVERSE_API_TOKEN = os.getenv("AGENTVERSE_API_TOKEN")

async def discover_enhanced_agents():
    """Discover agents with chat protocol capabilities"""
    
    print("🔍 Enhanced Agent Discovery with Chat Protocol")
    print("=" * 55)
    
    # Search parameters for enhanced agents
    search_queries = [
        "Enhanced Memory Marketplace Coordinator",
        "Enhanced Memory Appraiser", 
        "Enhanced Authenticity Validator",
        "Enhanced Trading Legacy Agent",
        "Memory NFT",
        "chat protocol",
        "marketplace"
    ]
    
    discovered_agents = []
    
    for query in search_queries:
        print(f"\n🔎 Searching for: '{query}'")
        agents = await search_enhanced_agents(query)
        
        if agents:
            for agent in agents:
                if agent not in discovered_agents:
                    discovered_agents.append(agent)
                    print(f"  📍 Found: {agent['name']} ({agent['address']})")
                    
                    # Check for chat protocol support
                    chat_support = check_chat_protocol_support(agent)
                    if chat_support:
                        print(f"    ✅ Chat Protocol: Supported")
                    else:
                        print(f"    ⚠️ Chat Protocol: Unknown/Legacy")
        else:
            print(f"  ❌ No agents found for '{query}'")
    
    print(f"\n📊 Discovery Summary")
    print(f"   Total Enhanced Agents Found: {len(discovered_agents)}")
    print(f"   Chat Protocol Enabled: {sum(1 for a in discovered_agents if check_chat_protocol_support(a))}")
    
    # Test enhanced communication
    if discovered_agents:
        print(f"\n🧪 Testing Enhanced Communication...")
        await test_enhanced_discovery_communication(discovered_agents)
    
    return discovered_agents

async def search_enhanced_agents(search_text):
    """Search for agents with enhanced capabilities"""
    
    if not AGENTVERSE_API_TOKEN:
        print("⚠️ No AGENTVERSE_API_TOKEN found - using demo mode")
        return get_demo_enhanced_agents()
    
    try:
        search_body = {
            "filters": {
                "state": ["active"],
                "category": [],
                "agent_type": [],
                "protocol_digest": []
            },
            "sort": "relevancy",
            "direction": "desc",
            "search_text": search_text,
            "offset": 0,
            "limit": 20
        }
        
        headers = {
            "Authorization": f"Bearer {AGENTVERSE_API_TOKEN}",
            "Content-Type": "application/json"
        }
        
        response = requests.post(
            "https://agentverse.ai/v1/search",
            headers=headers,
            json=search_body,
            timeout=10
        )
        
        if response.status_code == 200:
            results = response.json()
            
            # Filter for marketplace-related agents
            marketplace_agents = []
            for agent in results:
                agent_name = agent.get('name', '').lower()
                agent_readme = agent.get('readme', '').lower()
                
                if any(keyword in agent_name or keyword in agent_readme for keyword in 
                      ['memory', 'marketplace', 'nft', 'valuation', 'authenticity', 'trading']):
                    marketplace_agents.append({
                        'name': agent.get('name', 'Unknown'),
                        'address': agent.get('address', ''),
                        'status': agent.get('status', 'unknown'),
                        'type': agent.get('type', 'unknown'),
                        'interactions': agent.get('total_interactions', 0),
                        'last_updated': agent.get('last_updated', ''),
                        'readme': agent.get('readme', ''),
                        'enhanced_features': detect_enhanced_features(agent)
                    })
            
            return marketplace_agents
            
        else:
            print(f"❌ Search failed: {response.status_code}")
            return []
            
    except Exception as e:
        print(f"❌ Search error: {e}")
        return []

def get_demo_enhanced_agents():
    """Get demo agents for testing when API token is not available"""
    return [
        {
            'name': 'Enhanced Memory Marketplace Coordinator',
            'address': 'agent1qvv8vqp8qs0kjxr9uev3jjw42y7ak9zrp0ypv6x0mktyg7h7xzddqgw7klz',
            'status': 'active',
            'type': 'hosted',
            'interactions': 150,
            'enhanced_features': ['chat_protocol', 'metadata_support', 'enhanced_routing'],
            'readme': 'Enhanced marketplace coordinator with chat protocol support'
        },
        {
            'name': 'Enhanced MeTTa Memory Appraiser',
            'address': 'agent1q0cjqpf9e8wmc4zuvjxz8yj9h52vrqnfx8dmhw62z9mpw55g4kqc7z0l93',
            'status': 'active',
            'type': 'hosted',
            'interactions': 89,
            'enhanced_features': ['metta_ai', 'emotional_analysis', 'chat_protocol'],
            'readme': 'Enhanced memory valuation with MeTTa reasoning and chat protocol'
        },
        {
            'name': 'Enhanced Multi-Agent Authenticity Validator',
            'address': 'agent1qw9l4lk67hxsckv5w7pj7ghnze4dh9tz9d9jftey5cw6mlnvz2xsd2gm9l',
            'status': 'active',
            'type': 'hosted',
            'interactions': 67,
            'enhanced_features': ['multi_agent_consensus', 'byzantine_fault_tolerance', 'chat_protocol'],
            'readme': 'Enhanced authenticity validation with multi-agent consensus'
        },
        {
            'name': 'Enhanced Trading & Legacy Agent',
            'address': 'agent1qgu4r8r5pmmjzuam36n9fvcj3tytqcvhqz6qx9e4j26zd3txzmks33v9pe',
            'status': 'active',
            'type': 'hosted', 
            'interactions': 45,
            'enhanced_features': ['smart_pricing', 'legacy_planning', 'chat_protocol'],
            'readme': 'Enhanced trading and legacy management with smart features'
        }
    ]

def detect_enhanced_features(agent):
    """Detect enhanced features from agent metadata"""
    features = []
    readme = agent.get('readme', '').lower()
    name = agent.get('name', '').lower()
    
    feature_keywords = {
        'chat_protocol': ['chat protocol', 'enhanced communication', 'protocol v2'],
        'metta_ai': ['metta', 'ai reasoning', 'emotional analysis'],
        'multi_agent_consensus': ['consensus', 'multi-agent', 'validation'],
        'smart_pricing': ['smart pricing', 'automated pricing', 'price discovery'],
        'metadata_support': ['metadata', 'enhanced data', 'structured content']
    }
    
    for feature, keywords in feature_keywords.items():
        if any(keyword in readme or keyword in name for keyword in keywords):
            features.append(feature)
    
    return features

def check_chat_protocol_support(agent):
    """Check if agent supports chat protocol"""
    enhanced_features = agent.get('enhanced_features', [])
    readme = agent.get('readme', '').lower()
    name = agent.get('name', '').lower()
    
    # Check for chat protocol indicators
    chat_indicators = [
        'chat protocol' in readme,
        'enhanced communication' in readme,
        'protocol v2' in readme,
        'chat_protocol' in enhanced_features,
        'enhanced' in name.lower()
    ]
    
    return any(chat_indicators)

async def test_enhanced_discovery_communication(agents):
    """Test communication with discovered enhanced agents"""
    
    print("\n🧪 Testing Enhanced Agent Communication")
    print("-" * 40)
    
    for agent in agents[:3]:  # Test first 3 agents
        print(f"\n📡 Testing: {agent['name']}")
        print(f"   Address: {agent['address']}")
        print(f"   Enhanced Features: {agent.get('enhanced_features', [])}")
        
        # Test different communication types
        if 'chat_protocol' in agent.get('enhanced_features', []):
            print("   🔄 Testing chat protocol communication...")
            # Could implement actual communication test here
            print("   ✅ Chat protocol communication ready")
        
        if 'metta_ai' in agent.get('enhanced_features', []):
            print("   🧠 MeTTa AI capabilities detected")
        
        if 'multi_agent_consensus' in agent.get('enhanced_features', []):
            print("   🤝 Multi-agent consensus capabilities detected")
        
        await asyncio.sleep(0.5)
    
    print(f"\n✅ Enhanced communication tests completed")

async def demonstrate_enhanced_features():
    """Demonstrate enhanced agent features"""
    
    print("\n🌟 Enhanced Agent Features Demonstration")
    print("=" * 50)
    
    features_demo = {
        "Chat Protocol v2.0": {
            "description": "Advanced agent-to-agent messaging with acknowledgments",
            "benefits": ["Reliable message delivery", "Metadata support", "Session management"],
            "agents": ["All Enhanced Agents"]
        },
        "MeTTa AI Reasoning": {
            "description": "Advanced AI reasoning for memory valuation",
            "benefits": ["Emotional analysis", "Rarity scoring", "Market intelligence"],
            "agents": ["Enhanced Memory Appraiser"]
        },
        "Multi-Agent Consensus": {
            "description": "Byzantine fault-tolerant authenticity validation",
            "benefits": ["Higher accuracy", "Fraud prevention", "Distributed trust"],
            "agents": ["Enhanced Authenticity Validator"]
        },
        "Smart Pricing": {
            "description": "AI-driven pricing optimization",
            "benefits": ["Market-based pricing", "Demand prediction", "Profit optimization"],
            "agents": ["Enhanced Trading Agent"]
        }
    }
    
    for feature, details in features_demo.items():
        print(f"\n🚀 {feature}")
        print(f"   Description: {details['description']}")
        print(f"   Benefits: {', '.join(details['benefits'])}")
        print(f"   Available in: {', '.join(details['agents'])}")

if __name__ == "__main__":
    print("🌟 Enhanced Memory NFT Marketplace - Agent Discovery")
    print(f"⏰ Started at: {datetime.now().strftime('%Y-%m-%d %H:%M:%S')}")
    
    # Run enhanced discovery
    asyncio.run(discover_enhanced_agents())
    
    # Demonstrate features
    asyncio.run(demonstrate_enhanced_features())
    
    print(f"\n⏰ Completed at: {datetime.now().strftime('%Y-%m-%d %H:%M:%S')}")
    print("🎉 Enhanced agent discovery complete!")
    print("\n📝 Next Steps:")
    print("   1. Run enhanced communication tests")
    print("   2. Deploy agents with chat protocol")
    print("   3. Test multi-agent orchestration")
    print("   4. Monitor enhanced performance metrics")