decentralized-memory-marketplace/requirements.txt
uagents>=0.12.0
python-dotenv>=1.0.0
fetchai>=0.1.42
fastapi>=0.104.0
uvicorn>=0.24.0
scikit-learn>=1.3.0